Create lemon_count
This is a simple implementation of openCV techniques to detect no. of lemons in an image. Read the comments in the code for further understanding. The code is applicable to simplified images and might show faulty results for complex images. For heap of lemons, the value shown is a rough estimate of no. of lemons on the surface.